Use the latest Python 3.6 in the python3 image
Debian Stretch (which is the current stable and used as the base image
for `distroless`) doesn't have Python 3.6, nor will it.  Instead, we
grab the python3.6 packages from Debian Buster, which is the current
testing branch.

This falls heavily into the Works For Me category.  Since these images
are minimalist, using a package from `testing` shouldn't conflict with
other bits in the image, but my `dpkg`-fu is sufficiently weak that I
can't guarantee that.  It'd be nice to have somebody take a deeper look.