python: Add dash so os.system and Popen.subprocess work (#229)
The Python standard library includes os.system and Popen.subprocess which
can both take a string argument, which then executes that string in a
shell. Add dash so /bin/sh exists and the standard library works.

Argubly using /bin/sh in this way is dangerous and deprecated, but parts of
the standard library itself rely on this (e.g. ctypes). I suspect there are
many Python programs "in the wild" that do as well. If the goal of the
default base image for rules_docker and distroless is to reliably run any
Python program, it should support this, since this is part of the standard
library. The additional ~200 kiB seems worth the additional compatibility.

Found while investigating
https://github.com/GoogleContainerTools/distroless/issues/150