python: add libc-bin and ld.so.cache so ctypes works (#228)
Third-party packages like gunicorn and monotonic assume that the standard
library function ctypes.util.find_library can find 'c' and 'rt', which are
parts of glibc. To do this, they execute '/sbin/ldconfig -p'. To make this
work, install the binary in the libc-bin deb and generate /etc/ld.so.cache.

Ideally, we would generate /etc/ld.so.cache from the container at build
time. However, that would require users to have Docker. Instead, check in a
pre-generated binary, and add a test that checks that it equals the
generated version. This allows end users to build python images without
Docker, but developers working on Distroless will have a check that it is
up to date.

Fixes https://github.com/GoogleContainerTools/distroless/issues/150