WORKSPACE: Update libc-bin to latest released version (security) (#251)
Commit b35d94bc3d added libc-bin to the Python image. Since dpkg_parser
does not parse and compare versions, it is getting the most recent
version from debian_security, which is an older pre-release security fix,
rather than the released version. To fix it, ask for the right version.
A similar issue was fixed for libc6 itself in 6f5345dd25.

Updates libc-bin from 2.24-11+deb9u1 to 2.24-11+deb9u3

Found by the Google Cloud container security scanner.