WORKSPACE: Latest Debian snapshot. Update libc6 to
Update libc and libc2 from 2.24-11+deb9u3 -> 2.24-11+deb9u4. This is a
"regular" update, and not a security update, so we still need the hack to
force the latest version to be resolved instead of the older version in
the stretch-security release.

However, this does fix some low priority CVEs that are reported by the
Google Cloud container vulnerability scanner. See:
https://metadata.ftp-master.debian.org/changelogs//main/g/glibc/glibc_2.24-11+deb9u4_changelog

This also updates the following packages to their latest versions:

base-files
curl
libc-bin
libc6
libcurl3
libldap-2.4-2
libldap-common
libsqlite3-0
openjdk-11-jre-headless
tzdata