base: Make -debian10 images default to nonroot user
Running as a non-root user inside a container provides some additional
security. For example, the runc container escape would not work for
non-root users:

https://aws.amazon.com/blogs/compute/anatomy-of-cve-2019-5736-a-runc-container-escape/

The existing Distroless images need to default to a root user for
compatibility. However, the introduction of -debian10 images gives us
an opportunity to break backwards compatibility, since users will need
to opt-in to the update anyway.