python3 debian10: Install full version of distutils for virtualenv
With Debian 10 buster, the python3 package only includes a partial
version of the distutils package. This causes virtualenv to create
a somewhat broken virtualenv, which causes the following import
error when attempting to use it. This fixes the error by installing
the full version from the standard library. Add a test to ensure
it continues to work.

Fixes the following error:

ImportError: cannot import name 'dist' from 'distutils'