Adds libuuid1 to the Java distroless images.
This issue: https://github.com/GoogleContainerTools/distroless/issues/321 has reappeared in the Debian 10-based images.

After running with similar debugging flags to the original issue I managed to track it down to a missing `libuuid1` dependency under Debian 10.

This change will also add the dependency to the Debian 9-based image where it is not required but also has no effect.