WORKSPACE: Add buster-updates to debian10 to update ca-certificates (#527)
This updates ca-certificates to 20200601~deb10u1, which was added to
buster-updates. This may cause problems when packages are added to
buster-updates, but then updated again into the stable point releases.
For example, tzdata version 2019c-0+deb10u1 was added to
buster-updates, then later tzdata version 2020a-0+deb10u1 was added
to buster for the 10.4 "point release". Since dpkg_list takes the first
package it finds, it finds the older tzdata version.

We should eventually fix the dpkg_list function to actually attempt to
order by version numbers to resolve this. For now, I've manually pinned
the tzdata version to ensure it gets the latest version.

Co-authored-by: Chanseok Oh <chanseok@google.com>