Update to Bazel 3.2.0; Update all Bazel dependencies (#531)
* Update to Bazel 3.2.0; Update all Bazel dependencies

This updates basically all the tools used to develop distroless to
the latest versions. This doesn't seem to change anything, but some
of these things were quite old. Some minor tweaks were needed to get
the build to work. Some notable changes:

* We don't seem to need --host_force_python=PY2 unless the build
  environment has a broken version of Python3. Using the default
  (which I think is Python3 in recent versions of Bazel) seems to
  work. We should be using that now, since Python2 is now deprecated.

* Run CI on Ubuntu 18.04 instead of 14.04. This is a fairly recent
  LTS release.

* generate_ldconfig_cache.py: Either Bazel or the container rules
changed the path to the script that loads the image. A better
fix is probably to use the new container_run rules.

* .travis.yml: Remove a bunch of config; this should all be brought in
by Bazel I think?

* .travis.yml: Remove sudo: required; no longer supported