python: Make tests faster: Import modules in a single container (#536)
* python: Make tests faster: Import modules in a single container

The previous version of the test imported a number of standard
library modules in seperate tests. This imports all the modules
in a single container. This is much faster.

Before:

    //experimental/python3:python3_debian10_test  PASSED in 306.6s

After:

    //experimental/python3:python3_debian10_test  PASSED in 14.8s

I also checked that this seems to find all the package errors.
I removed a few packages from BUILD and the expected test failed.
E.g. removed libbz2 makes "import bz2" fail.

This fixes:
https://github.com/GoogleContainerTools/distroless/issues/534

It is an alternative to:
https://github.com/GoogleContainerTools/distroless/pull/535

* code review improvements