package_manager dpkg_parser: Use /usr/bin/env python3
The dpkg_parser.py file imports the lzma system package, which only
exists in Python 3. On my system (Debian 10), the default "python"
is still Python 2. As a result, the built dpkg_parser.par does not
execute. Update the interpreter to use "python3" explicitly.

The python_version attribute does nothing at the moment, but it seems
to me that is likely a bug in the subpar bazel rules. Let's set it
since that seems to be the documented way to change the Python
version for a py_binary.