experimental/python3: Add Debian 11 support
This edits the tests a bit to ensure the container works. I have
not tested this, beyond getting the tests to pass. However, those
tests do check that all the standard library modules work, so
this is at least a good starting point.

WORKSPACE: sort the packages in the java and python sections to make
it easier to see differences between two lists.