Refactor the cosign script.
This refactors the cosign script for distroless, so that the arguments are passed through from `cloudbuild.yaml`, and shifts where `-key ...` is passed to be in `cloudbuild.yaml`.

This is preparation to start *also* signing the distroless images with a "keyless" flow, once https://github.com/sigstore/cosign/pull/652 lands and cosign 1.2 cuts.