README.md: Update to use Debian 11; Other minor doc updates
Debian 11 is now the "default" Distroless operating system. Change
the examples to reference it explicitly. Adding the explicit OS
version is less likely to cause broken builds in the future, so it
seems like it is probably a good idea to recommend that.

Update the text about image sizes. It turns out that time zone info
is pretty big, so the smallest image is now a couple of MiB.