Update Bazel 3.4.1 -> 4.2.1 (LTS release)
This updates Bazel to the most recent long term support Bazel
release. This appears to work for me locally. However, I am unable
to test the Github Actions or Cloud Build changes, since I don't have
permission. This doesn't seem to require any changes to this
repository. It seems useful to make sure we are building with an LTS
version of Bazel?

The big change that is necessary is finding a public container image
that contains this version of Bazel. The marketplace images are no
longer being updated. The Google Cloud Build images are, but don't
have Bazel version tags. I manually executed the following docker
commands to get the latest version, and confirm this is using Bazel
4.2.1:

    docker pull gcr.io/cloud-builders/bazel
    docker run --rm -ti gcr.io/cloud-builders/bazel version
    docker image inspect gcr.io/cloud-builders/bazel --format '{{.Id}}'

See: https://github.com/GoogleCloudPlatform/cloud-builders/issues/667