Disable writing control info for node archives
This was messing with scanners and they were incorrectly
matching our installed nodejs binaries against the debian
security database which did not have matching entries for these
latest versions. Instead keep writing sboms and suggest to users
of scanners to use a scanner that is aware of nodejs binary
releases.

Signed-off-by: Appu Goundan <appu@google.com>